# Improve `Set` initialization performance